+++ src/Presentation/Nop.Web/Administration/Controllers/CheckoutAttributeController.cs
@@ -12,6 +12,7 @@ using Nop.Services.Orders;
 using Nop.Services.Tax;
 using Nop.Web.Framework.Controllers;
 using Telerik.Web.Mvc;
+using Nop.Services.Security;
 
 namespace Nop.Admin.Controllers
 {
@@ -31,6 +32,7 @@ namespace Nop.Admin.Controllers
         private readonly IMeasureService _measureService;
         private readonly MeasureSettings _measureSettings;
         private readonly ICustomerActivityService _customerActivityService;
+        private readonly IPermissionService _permissionService;
 
         #endregion
 
@@ -41,7 +43,8 @@ namespace Nop.Admin.Controllers
             ILocalizationService localizationService, ITaxCategoryService taxCategoryService,
             IWorkContext workContext, ICurrencyService currencyService, 
             ICustomerActivityService customerActivityService, CurrencySettings currencySettings,
-            IMeasureService measureService, MeasureSettings measureSettings)
+            IMeasureService measureService, MeasureSettings measureSettings,
+            IPermissionService permissionService)
         {
             this._checkoutAttributeService = checkoutAttributeService;
             this._languageService = languageService;
@@ -54,6 +57,7 @@ namespace Nop.Admin.Controllers
             this._currencySettings = currencySettings;
             this._measureService = measureService;
             this._measureSettings = measureSettings;
+            this._permissionService = permissionService;
         }
 
         #endregion
@@ -114,6 +118,9 @@ namespace Nop.Admin.Controllers
 
         public ActionResult List()
         {
+            if (!_permissionService.Authorize(StandardPermissionProvider.ManageCatalog))
+                return AccessDeniedView();
+
             var checkoutAttributes = _checkoutAttributeService.GetAllCheckoutAttributes(false);
             var gridModel = new GridModel<CheckoutAttributeModel>
             {
@@ -131,6 +138,9 @@ namespace Nop.Admin.Controllers
         [HttpPost, GridAction(EnableCustomBinding = true)]
         public ActionResult List(GridCommand command)
         {
+            if (!_permissionService.Authorize(StandardPermissionProvider.ManageCatalog))
+                return AccessDeniedView();
+
             var checkoutAttributes = _checkoutAttributeService.GetAllCheckoutAttributes(false);
             var gridModel = new GridModel<CheckoutAttributeModel>
             {
@@ -151,6 +161,9 @@ namespace Nop.Admin.Controllers
         //create
         public ActionResult Create()
         {
+            if (!_permissionService.Authorize(StandardPermissionProvider.ManageCatalog))
+                return AccessDeniedView();
+
             var model = new CheckoutAttributeModel();
             //locales
             AddLocales(_languageService, model.Locales);
@@ -161,6 +174,9 @@ namespace Nop.Admin.Controllers
         [HttpPost, FormValueExists("save", "save-continue", "continueEditing")]
         public ActionResult Create(CheckoutAttributeModel model, bool continueEditing)
         {
+            if (!_permissionService.Authorize(StandardPermissionProvider.ManageCatalog))
+                return AccessDeniedView();
+
             if (ModelState.IsValid)
             {
                 var checkoutAttribute = model.ToEntity();
@@ -182,6 +198,9 @@ namespace Nop.Admin.Controllers
         //edit
         public ActionResult Edit(int id)
         {
+            if (!_permissionService.Authorize(StandardPermissionProvider.ManageCatalog))
+                return AccessDeniedView();
+
             var checkoutAttribute = _checkoutAttributeService.GetCheckoutAttributeById(id);
             if (checkoutAttribute == null)
                 throw new ArgumentException("No checkout attribute found with the specified id", "id");
@@ -200,6 +219,9 @@ namespace Nop.Admin.Controllers
         [HttpPost, FormValueExists("save", "save-continue", "continueEditing")]
         public ActionResult Edit(CheckoutAttributeModel model, bool continueEditing)
         {
+            if (!_permissionService.Authorize(StandardPermissionProvider.ManageCatalog))
+                return AccessDeniedView();
+
             var checkoutAttribute = _checkoutAttributeService.GetCheckoutAttributeById(model.Id);
             if (checkoutAttribute == null)
                 throw new ArgumentException("No checkout attribute found with the specified id");
@@ -226,6 +248,9 @@ namespace Nop.Admin.Controllers
         [HttpPost, ActionName("Delete")]
         public ActionResult DeleteConfirmed(int id)
         {
+            if (!_permissionService.Authorize(StandardPermissionProvider.ManageCatalog))
+                return AccessDeniedView();
+
             var checkoutAttribute = _checkoutAttributeService.GetCheckoutAttributeById(id);
             _checkoutAttributeService.DeleteCheckoutAttribute(checkoutAttribute);
 
@@ -244,6 +269,9 @@ namespace Nop.Admin.Controllers
         [HttpPost, GridAction(EnableCustomBinding = true)]
         public ActionResult ValueList(int checkoutAttributeId, GridCommand command)
         {
+            if (!_permissionService.Authorize(StandardPermissionProvider.ManageCatalog))
+                return AccessDeniedView();
+
             var values = _checkoutAttributeService.GetCheckoutAttributeValues(checkoutAttributeId);
             var gridModel = new GridModel<CheckoutAttributeValueModel>
             {
@@ -268,6 +296,9 @@ namespace Nop.Admin.Controllers
         //create
         public ActionResult ValueCreatePopup(int checkoutAttributeId)
         {
+            if (!_permissionService.Authorize(StandardPermissionProvider.ManageCatalog))
+                return AccessDeniedView();
+
             var model = new CheckoutAttributeValueModel();
             model.CheckoutAttributeId = checkoutAttributeId;
             model.PrimaryStoreCurrencyCode = _currencyService.GetCurrencyById(_currencySettings.PrimaryStoreCurrencyId).CurrencyCode;
@@ -281,6 +312,9 @@ namespace Nop.Admin.Controllers
         [HttpPost]
         public ActionResult ValueCreatePopup(string btnId, CheckoutAttributeValueModel model)
         {
+            if (!_permissionService.Authorize(StandardPermissionProvider.ManageCatalog))
+                return AccessDeniedView();
+
             var checkoutAttribute = _checkoutAttributeService.GetCheckoutAttributeById(model.CheckoutAttributeId);
             if (checkoutAttribute == null)
                 throw new ArgumentException("No checkout attribute found with the specified id");
@@ -307,6 +341,9 @@ namespace Nop.Admin.Controllers
         //edit
         public ActionResult ValueEditPopup(int id)
         {
+            if (!_permissionService.Authorize(StandardPermissionProvider.ManageCatalog))
+                return AccessDeniedView();
+
             var cav = _checkoutAttributeService.GetCheckoutAttributeValueById(id);
             if (cav == null)
                 throw new ArgumentException("No checkout attribute value found with the specified id", "id");
@@ -326,6 +363,9 @@ namespace Nop.Admin.Controllers
         [HttpPost]
         public ActionResult ValueEditPopup(string btnId, CheckoutAttributeValueModel model)
         {
+            if (!_permissionService.Authorize(StandardPermissionProvider.ManageCatalog))
+                return AccessDeniedView();
+
             var cav = _checkoutAttributeService.GetCheckoutAttributeValueById(model.Id);
             if (cav == null)
                 throw new ArgumentException("No checkout attribute value found with the specified id");
@@ -352,6 +392,9 @@ namespace Nop.Admin.Controllers
         [GridAction(EnableCustomBinding = true)]
         public ActionResult ValueDelete(int valueId, int checkoutAttributeId, GridCommand command)
         {
+            if (!_permissionService.Authorize(StandardPermissionProvider.ManageCatalog))
+                return AccessDeniedView();
+
             var cav = _checkoutAttributeService.GetCheckoutAttributeValueById(valueId);
             _checkoutAttributeService.DeleteCheckoutAttributeValue(cav);
